Move authorized namespace fetch endpoint to a separate user API
